feat(#163): Add CLI commands for git and GitHub operations with AI integration
