+++ .gitignore
@@ -1,5 +1,5 @@
 .aider*
-aidy
+/aidy
 .idea
 coverage.out
 .aidy
