+++ internal/ai/ai.go
