+++ internal/ai/ai_test.go
