+++ internal/ai/mockai.go
