+++ internal/ai/mockai_test.go
