+++ internal/ai/openai.go
