+++ internal/ai/openai_test.go
