+++ internal/ai/prompts.go
