+++ internal/config/aiderconfig.go
