+++ internal/config/aiderconfig_test.go
