+++ internal/github/github_test.go
@@ -7,8 +7,8 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/volodya-lombrozo/aidy/cache"
-	"github.com/volodya-lombrozo/aidy/git"
+	"github.com/volodya-lombrozo/aidy/internal/cache"
+	"github.com/volodya-lombrozo/aidy/internal/git"
 )
 
 const JsonIssue = `{
